Zorgen dat Orders goed geïntegreerd zijn met Mollie-orders. Backend orders ontbreekt grotendeels.
Former-commit-id: 139045f47ced2e1fe6f82fd8accb362e6abb3c2f

+++ app/Store/Order.php
@@ -67,17 +67,19 @@
         }
 
         /**
+         * @param bool $deleteOrder
+         *
          * @throws \Throwable
          */
-        public function undoOrder() {
+        public function undoOrder($deleteOrder = false) {
             foreach ($this->items as $item) {
                 $stock = $item->item;
                 if($stock !== null) {
                     $stock->in_stock = $stock->in_stock + $item->amount;
                     $stock->saveOrFail();
                 }
-                $item->delete();
+                if($deleteOrder) $item->delete();
             }
-            $this->delete();
+            if($deleteOrder) $this->delete();
         }
     }
